Move all content from DicoveryScreen to ConnectedAppsBottomSheet
